103 - Fixing build failure: moves layout file into components
